Added - Cinemachine Package
This commit just adds the Cinemachine package which will allow me to program better cameras and adjust them much easier than with the standard Unity Camera System.

Next, work on implementing a script to allow the player to move the golf club back and forward for a more/less powerful shot.