Implemented - Golf Club Central Rotation
This commit implements the rotation and positioning script for the Golf Club. It's not exactly how I wanted it to be as I have had to prevent player's from moving the camera along the Y axis but it's working perfectly on the X.

Next, implement the actual pulling back of the golf club to set up a shot.